Move Loader.PathResolver to Package.Loader.PathResolver
